Initial breakout of inventory logic into separate packages
