+++ image/manifest/manifest.go
@@ -26,6 +26,7 @@ import (
 	"golang.org/x/xerrors"
 
 	reg "sigs.k8s.io/promo-tools/v3/internal/legacy/dockerregistry"
+	"sigs.k8s.io/promo-tools/v3/internal/legacy/dockerregistry/registry"
 	"sigs.k8s.io/promo-tools/v3/types/image"
 )
 
@@ -137,7 +138,7 @@ func Grow(
 	o *GrowOptions,
 ) error {
 	var err error
-	var riiCombined reg.RegInvImage
+	var riiCombined registry.RegInvImage
 
 	// (1) Scan the BaseDir and find the promoter manifest to modify.
 	manifest, err := Find(o)
@@ -170,7 +171,7 @@ func Grow(
 
 // Write writes images as YAML out to the expected path of the given
 // (thin) manifest.
-func Write(manifest reg.Manifest, rii reg.RegInvImage) error {
+func Write(manifest reg.Manifest, rii registry.RegInvImage) error {
 	// Chop off trailing "promoter-manifest.yaml".
 	p := path.Dir(manifest.Filepath)
 	// Get staging repo directory name as it is laid out in the thin manifest
@@ -183,7 +184,7 @@ func Write(manifest reg.Manifest, rii reg.RegInvImage) error {
 
 	// Write the file.
 	err := ioutil.WriteFile(
-		imagesPath, []byte(rii.ToYAML(reg.YamlMarshalingOpts{})), 0o644)
+		imagesPath, []byte(rii.ToYAML(registry.YamlMarshalingOpts{})), 0o644)
 	return err
 }
 
@@ -211,7 +212,7 @@ func Find(o *GrowOptions) (reg.Manifest, error) {
 // inject into the "images.yaml" of a thin manifest.
 func ReadStagingRepo(
 	o *GrowOptions,
-) (reg.RegInvImage, error) {
+) (registry.RegInvImage, error) {
 	stagingRepoRC := reg.RegistryContext{
 		Name: o.StagingRepo,
 	}
@@ -231,7 +232,7 @@ func ReadStagingRepo(
 		true,
 		false)
 	if err != nil {
-		return reg.RegInvImage{}, err
+		return registry.RegInvImage{}, err
 	}
 	sc.ReadRegistries(
 		[]reg.RegistryContext{stagingRepoRC},
@@ -245,7 +246,7 @@ func ReadStagingRepo(
 
 // ApplyFilters applies the filters in the options to whittle down the given
 // rii.
-func ApplyFilters(o *GrowOptions, rii reg.RegInvImage) (reg.RegInvImage, error) {
+func ApplyFilters(o *GrowOptions, rii registry.RegInvImage) (registry.RegInvImage, error) {
 	// If nothing to filter, short-circuit.
 	if len(rii) == 0 {
 		return rii, nil
@@ -270,7 +271,7 @@ func ApplyFilters(o *GrowOptions, rii reg.RegInvImage) (reg.RegInvImage, error) 
 	rii = ExcludeTags(rii, excludeTags)
 
 	if len(rii) == 0 {
-		return reg.RegInvImage{}, xerrors.New(
+		return registry.RegInvImage{}, xerrors.New(
 			"no images survived filtering; double-check your --filter_* flag(s) for typos",
 		)
 	}
@@ -280,8 +281,8 @@ func ApplyFilters(o *GrowOptions, rii reg.RegInvImage) (reg.RegInvImage, error) 
 
 // FilterByImages removes all images in RegInvImage that do not match the
 // filterImage.
-func FilterByImages(rii reg.RegInvImage, filterImages []image.Name) reg.RegInvImage {
-	filtered := make(reg.RegInvImage)
+func FilterByImages(rii registry.RegInvImage, filterImages []image.Name) registry.RegInvImage {
+	filtered := make(registry.RegInvImage)
 	for imageName, digestTags := range rii {
 		for _, filterImage := range filterImages {
 			if imageName == filterImage {
@@ -295,8 +296,8 @@ func FilterByImages(rii reg.RegInvImage, filterImages []image.Name) reg.RegInvIm
 // FilterByTags removes all images in RegInvImage that do not match the
 // filterTag.
 // TODO(manifest): Dedupe with `FilterByTag` in legacy/dockerregistry/inventory.go
-func FilterByTags(rii reg.RegInvImage, filterTags []image.Tag) reg.RegInvImage {
-	filtered := make(reg.RegInvImage)
+func FilterByTags(rii registry.RegInvImage, filterTags []image.Tag) registry.RegInvImage {
+	filtered := make(registry.RegInvImage)
 
 	for imageName, digestTags := range rii {
 		for digest, tags := range digestTags {
@@ -304,7 +305,7 @@ func FilterByTags(rii reg.RegInvImage, filterTags []image.Tag) reg.RegInvImage {
 				for _, filterTag := range filterTags {
 					if tag == filterTag {
 						if filtered[imageName] == nil {
-							filtered[imageName] = make(reg.DigestTags)
+							filtered[imageName] = make(registry.DigestTags)
 						}
 
 						filtered[imageName][digest] = append(
@@ -322,14 +323,14 @@ func FilterByTags(rii reg.RegInvImage, filterTags []image.Tag) reg.RegInvImage {
 
 // FilterByDigests removes all images in RegInvImage that do not match the
 // filterDigest.
-func FilterByDigests(rii reg.RegInvImage, filterDigests []image.Digest) reg.RegInvImage {
-	filtered := make(reg.RegInvImage)
+func FilterByDigests(rii registry.RegInvImage, filterDigests []image.Digest) registry.RegInvImage {
+	filtered := make(registry.RegInvImage)
 	for imageName, digestTags := range rii {
 		for digest, tags := range digestTags {
 			for _, filterDigest := range filterDigests {
 				if digest == filterDigest {
 					if filtered[imageName] == nil {
-						filtered[imageName] = make(reg.DigestTags)
+						filtered[imageName] = make(registry.DigestTags)
 					}
 					filtered[imageName][digest] = tags
 				}
@@ -341,8 +342,8 @@ func FilterByDigests(rii reg.RegInvImage, filterDigests []image.Digest) reg.RegI
 }
 
 // ExcludeTags removes tags in rii that match excludedTags.
-func ExcludeTags(rii reg.RegInvImage, excludedTags map[image.Tag]bool) reg.RegInvImage {
-	filtered := make(reg.RegInvImage)
+func ExcludeTags(rii registry.RegInvImage, excludedTags map[image.Tag]bool) registry.RegInvImage {
+	filtered := make(registry.RegInvImage)
 	for imageName, digestTags := range rii {
 		for digest, tags := range digestTags {
 			for _, tag := range tags {
@@ -350,7 +351,7 @@ func ExcludeTags(rii reg.RegInvImage, excludedTags map[image.Tag]bool) reg.RegIn
 					continue
 				}
 				if filtered[imageName] == nil {
-					filtered[imageName] = make(reg.DigestTags)
+					filtered[imageName] = make(registry.DigestTags)
 				}
 				filtered[imageName][digest] = append(
 					filtered[imageName][digest],
@@ -362,7 +363,7 @@ func ExcludeTags(rii reg.RegInvImage, excludedTags map[image.Tag]bool) reg.RegIn
 }
 
 // Union inject b's contents into a. However, it does so in a special way.
-func Union(a, b reg.RegInvImage) reg.RegInvImage {
+func Union(a, b registry.RegInvImage) registry.RegInvImage {
 	for imageName, digestTags := range b {
 		// If a does not have this image at all, then it's a simple
 		// injection.
@@ -379,7 +380,7 @@ func Union(a, b reg.RegInvImage) reg.RegInvImage {
 			}
 			// If c has the digest already, try to inject those tags in b that
 			// are not already in a.
-			tagSlice := reg.TagSlice{}
+			tagSlice := registry.TagSlice{}
 			for tag := range tags.Union(a[imageName][digest]) {
 				if tag == "latest" {
 					continue
