+++ internal/legacy/dockerregistry/inventory.go
@@ -24,10 +24,8 @@ import (
 	"io/ioutil"
 	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"regexp"
-	"runtime"
 	"sort"
 	"strings"
 	"sync"
@@ -44,6 +42,7 @@ import (
 	"github.com/sirupsen/logrus"
 	yaml "gopkg.in/yaml.v2"
 
+	"sigs.k8s.io/promo-tools/v3/internal/legacy/dockerregistry/registry"
 	"sigs.k8s.io/promo-tools/v3/internal/legacy/gcloud"
 	cipJson "sigs.k8s.io/promo-tools/v3/internal/legacy/json"
 	"sigs.k8s.io/promo-tools/v3/internal/legacy/reqcounter"
@@ -704,7 +703,7 @@ func (edge *PromotionEdge) VertexPropsFor(
 					p.PqinDigestMatch = true
 					// Both the digest and tag match what we wanted in the
 					// imageTag, so there are no extraneous tags to bother with.
-					p.OtherTags = TagSlice{}
+					p.OtherTags = registry.TagSlice{}
 				} else {
 					p.BadDigest = digest
 				}
@@ -1173,10 +1172,10 @@ func (sc *SyncContext) ReadRegistriesGGCR(
 					return errors.Wrap(err, "splitting repo and image name")
 				}
 				logrus.Infof("Registry: %s Image: %s Got: %s", regName, imageName, repo.Name())
-				digestTags := make(DigestTags)
+				digestTags := make(registry.DigestTags)
 				if tags != nil && tags.Manifests != nil {
 					for digest, manifest := range tags.Manifests {
-						tagSlice := TagSlice{}
+						tagSlice := registry.TagSlice{}
 						for _, tag := range manifest.Tags {
 							tagSlice = append(tagSlice, image.Tag(tag))
 						}
@@ -1196,7 +1195,7 @@ func (sc *SyncContext) ReadRegistriesGGCR(
 
 				mutex.Lock()
 				if _, ok := sc.Inv[regName]; !ok {
-					sc.Inv[regName] = make(RegInvImage)
+					sc.Inv[regName] = make(registry.RegInvImage)
 				}
 
 				if len(digestTags) > 0 {
@@ -1300,10 +1299,10 @@ func (sc *SyncContext) ReadRegistries(
 
 			// Process the current repo.
 			rName := req.RequestParams.(RegistryContext).Name
-			digestTags := make(DigestTags)
+			digestTags := make(registry.DigestTags)
 
 			for digest, mfestInfo := range tagsStruct.Manifests {
-				tagSlice := TagSlice{}
+				tagSlice := registry.TagSlice{}
 				for _, tag := range mfestInfo.Tags {
 					tagSlice = append(tagSlice, image.Tag(tag))
 				}
@@ -1333,7 +1332,7 @@ func (sc *SyncContext) ReadRegistries(
 					logrus.Fatal(err)
 				}
 
-				currentRepo := make(RegInvImage)
+				currentRepo := make(registry.RegInvImage)
 				currentRepo[imageName] = digestTags
 
 				mutex.Lock()
@@ -1495,15 +1494,15 @@ func (sc *SyncContext) ReadGCRManifestLists(
 
 // FilterByTag removes all images in RegInvImage that do not match the
 // filterTag.
-func FilterByTag(rii RegInvImage, filterTag string) RegInvImage {
-	filtered := make(RegInvImage)
+func FilterByTag(rii registry.RegInvImage, filterTag string) registry.RegInvImage {
+	filtered := make(registry.RegInvImage)
 
 	for imageName, digestTags := range rii {
 		for digest, tags := range digestTags {
 			for _, tag := range tags {
 				if string(tag) == filterTag {
 					if filtered[imageName] == nil {
-						filtered[imageName] = make(DigestTags)
+						filtered[imageName] = make(registry.DigestTags)
 					}
 
 					filtered[imageName][digest] = append(
@@ -1520,8 +1519,8 @@ func FilterByTag(rii RegInvImage, filterTag string) RegInvImage {
 
 // RemoveChildDigestEntries removes all tagless images in RegInvImage that are
 // referenced by ManifestLists in the Registries.
-func (sc *SyncContext) RemoveChildDigestEntries(rii RegInvImage) RegInvImage {
-	filtered := make(RegInvImage)
+func (sc *SyncContext) RemoveChildDigestEntries(rii registry.RegInvImage) registry.RegInvImage {
+	filtered := make(registry.RegInvImage)
 	for imageName, digestTags := range rii {
 		for digest, tagSlice := range digestTags {
 			_, hasParent := sc.ParentDigest[digest]
@@ -1533,7 +1532,7 @@ func (sc *SyncContext) RemoveChildDigestEntries(rii RegInvImage) RegInvImage {
 			}
 
 			if filtered[imageName] == nil {
-				filtered[imageName] = make(DigestTags)
+				filtered[imageName] = make(registry.DigestTags)
 			}
 
 			filtered[imageName][digest] = tagSlice
@@ -1799,116 +1798,6 @@ func ToPQIN(registryName image.Registry, imageName image.Name, tag image.Tag) st
 	return string(registryName) + "/" + string(imageName) + ":" + string(tag)
 }
 
-// ToSorted converts a RegInvImage type to a sorted structure.
-func (rii *RegInvImage) ToSorted() []ImageWithDigestSlice {
-	images := make([]ImageWithDigestSlice, 0)
-
-	for name, dmap := range *rii {
-		var digests []digest
-		for k, v := range dmap {
-			var tags []string
-			for _, tag := range v {
-				tags = append(tags, string(tag))
-			}
-
-			sort.Strings(tags)
-
-			digests = append(digests, digest{
-				hash: string(k),
-				tags: tags,
-			})
-		}
-		sort.Slice(digests, func(i, j int) bool {
-			return digests[i].hash < digests[j].hash
-		})
-
-		images = append(images, ImageWithDigestSlice{
-			name:    string(name),
-			digests: digests,
-		})
-	}
-
-	sort.Slice(images, func(i, j int) bool {
-		return images[i].name < images[j].name
-	})
-
-	return images
-}
-
-// ToYAML displays a RegInvImage as YAML, but with the map items sorted
-// alphabetically.
-func (rii *RegInvImage) ToYAML(o YamlMarshalingOpts) string {
-	images := rii.ToSorted()
-
-	var b strings.Builder
-	for _, image := range images {
-		fmt.Fprintf(&b, "- name: %s\n", image.name)
-		fmt.Fprintf(&b, "  dmap:\n")
-		for _, digestEntry := range image.digests {
-			if o.BareDigest {
-				fmt.Fprintf(&b, "    %s:", digestEntry.hash)
-			} else {
-				fmt.Fprintf(&b, "    %q:", digestEntry.hash)
-			}
-
-			switch len(digestEntry.tags) {
-			case 0:
-				fmt.Fprintf(&b, " []\n")
-			default:
-				if o.SplitTagsOverMultipleLines {
-					fmt.Fprintf(&b, "\n")
-					for _, tag := range digestEntry.tags {
-						fmt.Fprintf(&b, "    - %s\n", tag)
-					}
-				} else {
-					fmt.Fprintf(&b, " [")
-					for i, tag := range digestEntry.tags {
-						if i == len(digestEntry.tags)-1 {
-							fmt.Fprintf(&b, "%q", tag)
-						} else {
-							fmt.Fprintf(&b, "%q, ", tag)
-						}
-					}
-					fmt.Fprintf(&b, "]\n")
-				}
-			}
-		}
-	}
-
-	return b.String()
-}
-
-// ToCSV is like ToYAML, but instead of printing things in an indented
-// format, it prints one image on each line as a CSV. If there is a tag pointing
-// to the image, then it is printed next to the image on the same line.
-//
-// Example:
-// a@sha256:0000000000000000000000000000000000000000000000000000000000000000,a:1.0
-// a@sha256:0000000000000000000000000000000000000000000000000000000000000000,a:latest
-// b@sha256:1111111111111111111111111111111111111111111111111111111111111111,-
-func (rii *RegInvImage) ToCSV() string {
-	images := rii.ToSorted()
-
-	var b strings.Builder
-	for _, image := range images {
-		for _, digestEntry := range image.digests {
-			if len(digestEntry.tags) > 0 {
-				for _, tag := range digestEntry.tags {
-					fmt.Fprintf(&b, "%s@%s,%s:%s\n",
-						image.name,
-						digestEntry.hash,
-						image.name,
-						tag)
-				}
-			} else {
-				fmt.Fprintf(&b, "%s@%s,-\n", image.name, digestEntry.hash)
-			}
-		}
-	}
-
-	return b.String()
-}
-
 // ToLQIN converts a RegistryName and ImageName to form a loosely-qualified
 // image name (LQIN). Notice that it is missing tag information --- hence
 // "loosely-qualified".
@@ -2082,8 +1971,8 @@ func (sc *SyncContext) FilterPromotionEdges(
 func EdgesToRegInvImage(
 	edges map[PromotionEdge]interface{},
 	destRegistry string,
-) RegInvImage {
-	rii := make(RegInvImage)
+) registry.RegInvImage {
+	rii := make(registry.RegInvImage)
 
 	destRegistry = strings.TrimRight(destRegistry, "/")
 
@@ -2108,7 +1997,7 @@ func EdgesToRegInvImage(
 		}
 
 		if rii[image.Name(imgName)] == nil {
-			rii[image.Name(imgName)] = make(DigestTags)
+			rii[image.Name(imgName)] = make(registry.DigestTags)
 		}
 
 		digestTags := rii[image.Name(imgName)]
@@ -2117,7 +2006,7 @@ func EdgesToRegInvImage(
 				digestTags[edge.Digest],
 				edge.DstImageTag.Tag)
 		} else {
-			digestTags[edge.Digest] = TagSlice{}
+			digestTags[edge.Digest] = registry.TagSlice{}
 		}
 	}
 
@@ -2840,7 +2729,7 @@ func (payload *GCRPubSubPayload) String() string {
 
 // ParseSnapshot reads a given YAML snapshot from file.
 // TODO: Review/optimize/de-dupe (https://github.com/kubernetes-sigs/promo-tools/pull/351)
-func ParseSnapshot(pathToSnapshot string, images *[]ImageWithDigestSlice) error {
+func ParseSnapshot(pathToSnapshot string, images *[]registry.ImageWithDigestSlice) error {
 	f, err := os.Open(pathToSnapshot)
 	if err != nil {
 		return err
@@ -2849,7 +2738,7 @@ func ParseSnapshot(pathToSnapshot string, images *[]ImageWithDigestSlice) error 
 	defer f.Close()
 
 	reader := bufio.NewReader(f)
-	img := ImageWithDigestSlice{}
+	img := registry.ImageWithDigestSlice{}
 
 	// Converts output like `["v1.7.12", "v1.7.13-beta.0"]` into golang string array.
 	getTags := func(s string) []string {
@@ -2865,25 +2754,25 @@ func ParseSnapshot(pathToSnapshot string, images *[]ImageWithDigestSlice) error 
 	for output, _, err := reader.ReadLine(); err == nil; output, _, err = reader.ReadLine() {
 		line := string(output)
 		if strings.Contains(line, "- name:") {
-			if len(img.name) > 0 {
+			if len(img.Name) > 0 {
 				*images = append(*images, img)
 			}
 
 			imageName := line[8:]
-			img = ImageWithDigestSlice{
-				name:    imageName,
-				digests: make([]digest, 0),
+			img = registry.ImageWithDigestSlice{
+				Name:    imageName,
+				Digests: make([]registry.Digest, 0),
 			}
 		} else if strings.Contains(line, "sha256:") {
 			// Parse the line for sha256 and tags. The length and format of these
 			// lines are consistent, therefore it's safe to use fixed indices.
 			imageSha256 := line[5:76]
 			imageTags := getTags(line[79:])
-			img.digests = append(
-				img.digests,
-				digest{
-					hash: imageSha256,
-					tags: imageTags,
+			img.Digests = append(
+				img.Digests,
+				registry.Digest{
+					Hash: imageSha256,
+					Tags: imageTags,
 				},
 			)
 		}
@@ -2892,238 +2781,4 @@ func ParseSnapshot(pathToSnapshot string, images *[]ImageWithDigestSlice) error 
 	// Add the last image to the slice.
 	*images = append(*images, img)
 	return nil
-}
-
-// FilterParentImages filters the given ImageWithDigestSlice and only returns the images of
-// who contain manifest lists (aka: fat-manifests). This is determined by inspecting the
-// "mediaType" of every image manifest in the registry.
-//
-// TODO: Review/optimize/de-dupe (https://github.com/kubernetes-sigs/promo-tools/pull/351)
-func FilterParentImages(registry image.Registry, images *[]ImageWithDigestSlice) ([]ImageWithParentDigestSlice, error) {
-	// If an image is found to have this specific media type, it is a parent, and will be saved.
-	// All other images are not of interest.
-	mediaType := `"mediaType": "application/vnd.docker.distribution.manifest.list.v2+json"`
-	results := make([]ImageWithParentDigestSlice, 0)
-
-	// Split the registry into prefix and postfix
-	// For example: "us.gcr.io/k8s-artifacts-prod/addon-builder"
-	// prefix: "us.gcr.io"
-	// postfix: "/k8s-artifacts-prod/addon-builder"
-	firstSlash := strings.IndexByte(string(registry), '/')
-	registryPrefix := registry[:firstSlash]
-	registryPostfix := registry[firstSlash:]
-
-	// This is the number of goroutines that will be making curl requests over every image digest.
-	numWorkers := runtime.NumCPU() * 2
-	wg := new(sync.WaitGroup)
-	signal := make(chan ImageWithParentDigestSlice, 500)
-
-	filterImage := func(img ImageWithDigestSlice, c chan ImageWithParentDigestSlice) {
-		defer wg.Done()
-		imageLocation := fmt.Sprintf("%s/%s", registryPostfix, img.name)
-
-		// This is a standard endpoint all container registries must adopt. This will reveal
-		// information about the manifest for the particular image. The output is in JSON form,
-		// with a field "mediaType" that reveals if the given image is a parent image.
-		manifestEndpoint := fmt.Sprintf("https://%s/v2%s/manifests/", registryPrefix, imageLocation)
-		result := ImageWithParentDigestSlice{img.name, []parentDigest{}}
-		for _, digest := range img.digests {
-			var response string
-			numRetries := 8
-			imgEndpoint := manifestEndpoint + digest.hash
-			for numRetries > 0 {
-				out, err := exec.Command("curl", imgEndpoint).Output()
-				if err == nil {
-					response = string(out)
-					break
-				}
-				fmt.Println("something went wrong...")
-				fmt.Println("curl response:", err)
-				fmt.Println("retrying,", imgEndpoint)
-				numRetries--
-			}
-
-			if response == "" {
-				fmt.Println("Could not reach endpoint:", imgEndpoint)
-				continue
-			}
-
-			if !strings.Contains(response, mediaType) {
-				continue
-			}
-
-			// A new parent has been found.
-			parent := parentDigest{
-				hash:     digest.hash,
-				children: []string{},
-			}
-
-			// Parse all children by looking for digest fields within response.
-			children := strings.Split(response, "},")
-			for _, child := range children {
-				digestIdx := strings.LastIndex(child, `"digest":`)
-				if digestIdx != -1 {
-					childHash := child[digestIdx+11 : digestIdx+82]
-					parent.children = append(parent.children, childHash)
-				}
-			}
-
-			result.parentDigests = append(result.parentDigests, parent)
-		}
-
-		c <- result
-	}
-
-	fmt.Printf("Searching for parent images with %d workers...\n", numWorkers)
-	received := 0
-
-	// Engage all workers.
-	for i := 0; i < numWorkers; i++ {
-		img := (*images)[i]
-		go filterImage(img, signal)
-		wg.Add(1)
-	}
-
-	// When a worker finishes, create another.
-	for i := numWorkers; i < len(*images); i++ {
-		found := <-signal
-		received++
-		if len(found.parentDigests) > 0 {
-			results = append(results, found)
-			fmt.Println("FOUND parent:", found.Name)
-		}
-
-		img := (*images)[i]
-		go filterImage(img, signal)
-		wg.Add(1)
-	}
-
-	// Gather all results.
-	wg.Wait()
-	for found := range signal {
-		received++
-		if len(found.parentDigests) > 0 {
-			results = append(results, found)
-			fmt.Println("FOUND parent:", found.Name)
-		}
-
-		if received == len(*images) {
-			break
-		}
-	}
-
-	return results, nil
-}
-
-// TODO: Review/optimize/de-dupe (https://github.com/kubernetes-sigs/promo-tools/pull/351)
-func ValidateParentImages(registry image.Registry, images []ImageWithParentDigestSlice) {
-	numWorkers := runtime.NumCPU() * 2
-	wg := new(sync.WaitGroup)
-	signal := make(chan string, 500)
-
-	validateImg := func(image ImageWithParentDigestSlice, c chan string) {
-		defer wg.Done()
-		if IsParentImageValid(registry, image) {
-			fmt.Println("PASSED - ", image.Name)
-			c <- ""
-			return
-		}
-
-		fmt.Println("FAILED - ", image.Name)
-		c <- image.Name
-	}
-
-	fmt.Printf("Inspecting children of %d parents...\n", len(images))
-	received := 0
-
-	// Engage all workers.
-	for i := 0; i < numWorkers; i++ {
-		img := images[i]
-		go validateImg(img, signal)
-		wg.Add(1)
-	}
-
-	// When a worker finishes, create another.
-	for i := numWorkers; i < len(images); i++ {
-		found := <-signal
-		received++
-		if len(found) > 0 {
-			fmt.Println("FAILED ", found)
-		}
-
-		img := images[i]
-		go validateImg(img, signal)
-		wg.Add(1)
-	}
-
-	// Gather all results.
-	wg.Wait()
-	for found := range signal {
-		received++
-		if len(found) > 0 {
-			fmt.Println("FAILED ", found)
-		}
-
-		if received == len(images) {
-			break
-		}
-	}
-}
-
-// IsParentImageValid only returns true if all child images, from the parent's
-// manifest list, are from the same image location.
-// Example:
-//		VALID 	parent=gcr.io/foo/bar child=gcr.io/foo/bar
-// 		INVALID parent=gcr.io/foo/bar child=gcr.io/foo/bar/foo
-//
-// TODO: Review/optimize/de-dupe (https://github.com/kubernetes-sigs/promo-tools/pull/351)
-func IsParentImageValid(registry image.Registry, img ImageWithParentDigestSlice) bool {
-	// Split the registry into prefix and postfix
-	// For example: "us.gcr.io/k8s-artifacts-prod/addon-builder"
-	// prefix: "us.gcr.io"
-	// postfix: "/k8s-artifacts-prod/addon-builder"
-	firstSlash := strings.IndexByte(string(registry), '/')
-	registryPrefix := registry[:firstSlash]
-	registryPostfix := registry[firstSlash:]
-
-	imageLocation := fmt.Sprintf("%s/%s", registryPostfix, img.Name)
-	manifestEndpoint := fmt.Sprintf("https://%s/v2%s/manifests/", registryPrefix, imageLocation)
-	for _, parent := range img.parentDigests {
-		for _, childHash := range parent.children {
-			var response string
-
-			// This is a standard endpoint all container registries must adopt. This will reveal
-			// information about the manifest for the particular image. The output is in JSON form,
-			// with a field "mediaType" that reveals if the given image is a parent image.
-			imgEndpoint := manifestEndpoint + childHash
-			retries := 8
-
-			// Inspect the image. If it doesn't exist, we know the parent child relationship
-			// can't be assumed.
-			for retries > 0 {
-				out, err := exec.Command("curl", imgEndpoint).Output()
-				if err == nil {
-					response = string(out)
-					break
-				}
-
-				fmt.Println("trouble obtaining child manifest...")
-				fmt.Println("curl response:", err)
-				fmt.Println("retrying ", imgEndpoint)
-				retries--
-			}
-
-			if retries == 0 {
-				fmt.Println("Failed to reach ", imgEndpoint)
-				fmt.Println("skipping...")
-				continue
-			}
-
-			if !strings.Contains(response, "mediaType") {
-				return false
-			}
-		}
-	}
-
-	return true
 }
