+++ internal/legacy/dockerregistry/registry/set.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package inventory
+package registry
 
 import (
 	"sigs.k8s.io/promo-tools/v3/internal/legacy/container"
